LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-focus-navigation-checked-expected.txt





FAIL Focus the new checked radio in the focused radio group. assert_equals: expected Element node <input type="radio" name="radio" id="a"></input> but got Element node <input type="text" id="start"></input>
PASS Focusable radio can be focused even with a checked radio in the group.
FAIL Focus the checked radio in the group. assert_equals: expected Element node <input type="radio" name="radio"></input> but got Element node <input type="text" id="start"></input>
FAIL Focus the checked radio in the focused radio group. assert_false: expected false got true
FAIL When `checked` is false on a radio element, focus navigation target the radio group first element. assert_equals: expected Element node <input type="radio" name="radio" id="a"></input> but got Element node <input type="text" id="start"></input>
